game_opt_custom: Save custom players/difficulty/gsize/ai_id to cfg file

+++ src/game/game_new.h
@@ -7,8 +7,8 @@ struct game_new_options_s {
     uint32_t galaxy_seed;
     galaxy_size_t galaxy_size;
     difficulty_t difficulty;
-    uint8_t ai_id;
-    uint8_t players;
+    uint32_t ai_id;
+    uint32_t players;
     struct {
         char playername[EMPEROR_NAME_LEN];
         char homename[PLANET_NAME_LEN];
